fix: rename & strukturing folder
